logging: move logging code to a new file
In preparation to extend the logging use case move the existing code to
a new location.
