+++ util/logging.h
@@ -1,0 +1,10 @@
+/* SPDX-License-Identifier: GPL-2.0-or-later */
+
+#ifndef DEBUG_H_
+#define DEBUG_H_
+
+#include <stdbool.h>
+
+int map_log_level(int verbose, bool quiet);
+
+#endif // DEBUG_H_
