+++ util/meson.build
@@ -4,6 +4,7 @@ sources += [
   'util/argconfig.c',
   'util/base64.c',
   'util/crc32.c',
+  'util/logging.c',
   'util/mem.c',
   'util/suffix.c',
   'util/types.c',
